csv_load/requirements.txt
mysql-connector-python
configparser
